De-ratchify accac0d

+++ bot.py
@@ -1,5 +1,6 @@
-import sys
+import config
 import os
+import sys
 import signal
 import string
 import time
@@ -11,22 +12,38 @@ class StartServer(object):
         else:
             self.foreground = False
         self.pidfile = str(os.path.dirname(os.path.realpath(__file__))) + "/popm.pid"
+        self.fpid = None
+
+    def is_forked(self):
+        print "In thing"
+        if not self.fpid:
+            return False
+        else:
+            return True
 
     def is_process_running(self):
         f = open(self.pidfile, "r")
+        words = 0
         for line in f:
+            words += 1
             try:
                 pid = int(line)
             except:
                 return False
+        if words == 0:
+            return False
         try:
             os.kill(pid, 0)
         except OSError:
             return False
         else:
-            return True
+            sys.exit("POPM is already running under PID %s" % (pid))
 
     def run(self, argv):
+        import config
+        config.dbverify()
+        if os.path.isfile(self.pidfile) and '-h' not in str(sys.argv):
+            self.is_process_running()
         if '-h' in str(sys.argv):
             print ""
             print "POPM - Python Open Proxy Monitor"
@@ -39,43 +56,26 @@ class StartServer(object):
             print ""
             sys.exit()
         elif '-f' in str(sys.argv):
-            pid = str(os.getpid())
-            if os.path.isfile(self.pidfile):
-                if self.is_process_running():
-                    sys.exit("POPM is already running.")
-                else:
-                    file(self.pidfile, 'w').write(pid)
-            else:
-                file(self.pidfile, 'w').write(pid)
             print "Starting POPM...."
+            file(self.pidfile, 'w').write(str(os.getpid()))
             print "Now running in foreground mode"
             self.startprotocol()
         else:
-            fpid = os.fork()
-            if fpid != 0:
+            self.fpid = os.fork()
+            if self.fpid != 0:
                 print "Starting POPM...."
-                spid = str(fpid)
-                if os.path.isfile(self.pidfile):
-                    if self.is_process_running():
-                        print "POPM is already running."
-                        os.kill(fpid, signal.SIGTERM)
-                        os.kill(os.getpid(), signal.SIGTERM)
-                    else:
-                        file(self.pidfile, 'w').write(spid)
-                else:
-                    file(self.pidfile, 'w').write(spid)
-                print "Forking into background under PID %d" % (fpid)
+                file(self.pidfile, 'w').write(str(self.fpid))
+                print "Forking into background under PID %d" % (self.fpid)
                 sys.exit()
             self.startprotocol()
 
     def logger(self, importance, inputs):
-        import config
         if importance <= config.DEBUG_LEVEL:
             if not self.foreground:
                 print(inputs)
 
     def startprotocol(self):
-        import config
+        config.dbconnect()
         if config.PROTO.lower() == "p10server":
                 import p10server
                 boot_time = int(time.time())
